refactor: move `EVMProjectionExec` to its own file
